+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * **husky:** 修复pre-push脚本中换行符问题 ([c04501e](https://github.com/zjydipingxian/hooks-encode/commit/c04501e60778585426a170c7cd1002b56f32fbe3))
 * **husky:** 移除pre-push钩子中的自动推送 ([8026db4](https://github.com/zjydipingxian/hooks-encode/commit/8026db4b9336fae1e32ca10200996d9242503f04))
 * **husky:** 重构changelog更新流程 ([ec33bb5](https://github.com/zjydipingxian/hooks-encode/commit/ec33bb554bd8783700633d5921e1aa0aecf56ea2))
+* **husky:** 自动更新changelog并提交 ([68e5706](https://github.com/zjydipingxian/hooks-encode/commit/68e5706cf75e0c87f9401a67893aa8374b3d8272))
 
 
 ### Features